Update placements data type to be int so that values between 32767 and 2,147,483,647 cause the correct error message

+++ src/Sfa.Tl.Matching.Models/ViewModel/PlacementInformationViewModel.cs
@@ -14,7 +14,7 @@ namespace Sfa.Tl.Matching.Models.ViewModel
         [Required(ErrorMessage = "You must tell us whether the employer knows how many placements they want at this location")]
         public bool? PlacementsKnown { get; set; }
 
-        public short? Placements { get; set; }
+        public int? Placements { get; set; }
 
         public int RouteId { get; set; }
         public string Postcode { get; set; }
